// src/lib/api/generated/models/invite.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Marketplace API
 * API for Village Website & Marketplace
 * OpenAPI spec version: 1.0.0
 */
import type { InviteRole } from './inviteRole';
import type { InviteStatus } from './inviteStatus';

export interface Invite {
  id: string;
  email: string;
  orgId: string;
  /**
   * @maxLength 255
   * @nullable
   */
  code: string | null;
  role: InviteRole;
  status: InviteStatus;
  expiresAt: string;
  createdAt: string;
}