// src/lib/api/generated/models/orgInviteStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Marketplace API
 * API for Village Website & Marketplace
 * OpenAPI spec version: 1.0.0
 */

/**
 * The organization member invite status.
 */
export type OrgInviteStatus = typeof OrgInviteStatus[keyof typeof OrgInviteStatus];


export const OrgInviteStatus = {
  pending: 'pending',
  accepted: 'accepted',
  expired: 'expired',
} as const;